Rename MidasMemModel->FASED, remove SimpleLatencyPipe
